+++ tomcat/src/main/java/org/apache/coyote/http/vo/Url.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.domain;
+package org.apache.coyote.http.vo;
 
 public class Url {
