+++ tomcat/src/test/java/org/apache/coyote/domain/HttpRequestTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.util.List;
 import org.apache.coyote.http.HttpHeader;
 import org.apache.coyote.http.HttpMethod;
+import org.apache.coyote.http.vo.HttpRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
