+++ tomcat/src/test/java/org/apache/coyote/domain/QueryStringsTest.java
@@ -3,6 +3,7 @@ package org.apache.coyote.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import org.apache.coyote.http.vo.QueryStrings;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
